Crate drive2

Features

Handle the following Resources with ease from the central hub ...

about (get)

apps (get and list)

changes (get, list and watch)

channels (stop)

children (delete, get, insert and list)

comments (delete, get, insert, list, patch and update)

files (copy, delete, emptytrash, get, insert, list, patch, touch, trash, untrash, update and watch)

parents (delete, get, insert and list)

permissions (delete, get, getidforemail, insert, list, patch and update)

properties (delete, get, insert, list, patch and update)

realtime (get and update)

replies (delete, get, insert, list, patch and update)

revisions (delete, get, list, patch and update)

Everything else about the drive API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.files().watch(...).doit()
let r = hub.files().empty_trash(...).doit()
let r = hub.files().copy(...).doit()
let r = hub.files().list(...).doit()
let r = hub.files().delete(...).doit()
let r = hub.files().patch(...).doit()
let r = hub.files().update(...).doit()
let r = hub.files().insert(...).doit()
let r = hub.files().untrash(...).doit()
let r = hub.files().trash(...).doit()
let r = hub.files().touch(...).doit()
let r = hub.files().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate drive2;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: File = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.files().patch(&req, "fileId")
             .use_content_as_indexable_text(false)
             .update_viewed_date(true)
             .timed_text_track_name("gubergren")
             .timed_text_language("sadipscing")
             .set_modified_date(true)
             .remove_parents("ea")
             .pinned(false)
             .ocr_language("justo")
             .ocr(true)
             .new_revision(true)
             .convert(true)
             .add_parents("diam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

About
An item with user information and settings.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

AboutAdditionalRoleInfo
Information about supported additional roles per file type. The most specific type takes precedence.

This type is not used in any activity, and only used as part of another schema.

AboutAdditionalRoleInfoRoleSets
The supported additional roles per primary role.

This type is not used in any activity, and only used as part of another schema.

AboutExportFormats
The allowable export formats.

This type is not used in any activity, and only used as part of another schema.

AboutFeatures
List of additional features enabled on this account.

This type is not used in any activity, and only used as part of another schema.

AboutGetMethodBuilder
Gets the information about the current user along with Drive API settings

AboutImportFormats
The allowable import formats.

This type is not used in any activity, and only used as part of another schema.

AboutMaxUploadSizes
List of max upload sizes for each file type. The most specific type takes precedence.

This type is not used in any activity, and only used as part of another schema.

AboutMethodsBuilder
A builder providing access to all methods supported on about resources. It is not used directly, but through the Drive hub.

AboutQuotaBytesByService
The amount of storage quota used by different Google services.

This type is not used in any activity, and only used as part of another schema.

App
The apps resource provides a list of the apps that a user has installed, with information about each app's supported MIME types, file extensions, and other details.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

list (none)

AppGetMethodBuilder
Gets a specific app.

AppIcons
The various icons for the app.

This type is not used in any activity, and only used as part of another schema.

AppList
A list of third-party applications which the user has installed or given access to Google Drive.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

AppListMethodBuilder
Lists a user's installed apps.

AppMethodsBuilder
A builder providing access to all methods supported on app resources. It is not used directly, but through the Drive hub.

Change
Representation of a change to a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (none)

watch (none)

get (response)

ChangeGetMethodBuilder
Gets a specific change.

ChangeList
A list of changes for a user.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

ChangeListMethodBuilder
Lists the changes for a user.

ChangeMethodsBuilder
A builder providing access to all methods supported on change resources. It is not used directly, but through the Drive hub.

ChangeWatchMethodBuilder
Subscribe to changes for a user.

Channel
An notification channel used to watch for resource changes.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

watch (request|response)

watch (request|response)

stop (request)

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Drive hub.

ChannelStopMethodBuilder
Stop watching resources through this channel

ChildList
A list of children of a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

ChildReference
A reference to a folder's child.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert (request|response)

get (response)

ChildrenDeleteMethodBuilder
Removes a child from a folder.

ChildrenGetMethodBuilder
Gets a specific child reference.

ChildrenInsertMethodBuilder
Inserts a file into a folder.

ChildrenListMethodBuilder
Lists a folder's children.

ChildrenMethodsBuilder
A builder providing access to all methods supported on children resources. It is not used directly, but through the Drive hub.

Comment
A JSON representation of a comment on a file in Google Drive.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update (request|response)

insert (request|response)

delete (none)

patch (request|response)

get (response)

list (none)

CommentContext
The context of the file which is being commented on.

This type is not used in any activity, and only used as part of another schema.

CommentDeleteMethodBuilder
Deletes a comment.

CommentGetMethodBuilder
Gets a comment by ID.

CommentInsertMethodBuilder
Creates a new comment on the given file.

CommentList
A JSON representation of a list of comments on a file in Google Drive.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

CommentListMethodBuilder
Lists a file's comments.

CommentMethodsBuilder
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Drive hub.

CommentPatchMethodBuilder
Updates an existing comment. This method supports patch semantics.

CommentReply
A JSON representation of a reply to a comment on a file in Google Drive.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch (request|response)

insert (request|response)

get (response)

update (request|response)

CommentReplyList
A JSON representation of a list of replies to a comment on a file in Google Drive.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

CommentUpdateMethodBuilder
Updates an existing comment.

Drive
Central instance to access all Drive related resource activities

File
The metadata for a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

watch (none)

emptyTrash (none)

copy (request|response)

list (none)

delete (none)

patch (request|response)

update (request|response)

insert (request|response)

untrash (response)

trash (response)

touch (response)

get (response)

FileCopyMethodBuilder
Creates a copy of the specified file.

FileDeleteMethodBuilder
Permanently deletes a file by ID. Skips the trash. The currently authenticated user must own the file.

FileEmptyTrashMethodBuilder
Permanently deletes all of the user's trashed files.

FileGetMethodBuilder
Gets a file's metadata by ID.

FileImageMediaMetadata
Metadata about image media. This will only be present for image types, and its contents will depend on what can be parsed from the image content.

This type is not used in any activity, and only used as part of another schema.

FileImageMediaMetadataLocation
Geographic location information stored in the image.

This type is not used in any activity, and only used as part of another schema.

FileIndexableText
Indexable text attributes for the file (can only be written)

This type is not used in any activity, and only used as part of another schema.

FileInsertMethodBuilder
Insert a new file.

FileLabels
A group of labels for the file.

This type is not used in any activity, and only used as part of another schema.

FileList
A list of files.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

FileListMethodBuilder
Lists the user's files.

FileMethodsBuilder
A builder providing access to all methods supported on file resources. It is not used directly, but through the Drive hub.

FilePatchMethodBuilder
Updates file metadata and/or content. This method supports patch semantics.

FileThumbnail
Thumbnail for the file. Only accepted on upload and for files that are not already thumbnailed by Google.

This type is not used in any activity, and only used as part of another schema.

FileTouchMethodBuilder
Set the file's updated time to the current server time.

FileTrashMethodBuilder
Moves a file to the trash.

FileUntrashMethodBuilder
Restores a file from the trash.

FileUpdateMethodBuilder
Updates file metadata and/or content.

FileVideoMediaMetadata
Metadata about video media. This will only be present for video types.

This type is not used in any activity, and only used as part of another schema.

FileWatchMethodBuilder
Subscribe to changes on a file

ParentDeleteMethodBuilder
Removes a parent from a file.

ParentGetMethodBuilder
Gets a specific parent reference.

ParentInsertMethodBuilder
Adds a parent folder for a file.

ParentList
A list of a file's parents.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

ParentListMethodBuilder
Lists a file's parents.

ParentMethodsBuilder
A builder providing access to all methods supported on parent resources. It is not used directly, but through the Drive hub.

ParentReference
A reference to a file's parent.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get (response)

insert (request|response)

Permission
A permission for a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch (request|response)

list (none)

insert (request|response)

get (response)

getIdForEmail (none)

delete (none)

update (request|response)

PermissionDeleteMethodBuilder
Deletes a permission from a file.

PermissionGetIdForEmailMethodBuilder
Returns the permission ID for an email address.

PermissionGetMethodBuilder
Gets a permission by ID.

PermissionId
An ID for a user or group as seen in Permission items.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

getIdForEmail (response)

PermissionInsertMethodBuilder
Inserts a permission for a file.

PermissionList
A list of permissions associated with a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

PermissionListMethodBuilder
Lists a file's permissions.

PermissionMethodsBuilder
A builder providing access to all methods supported on permission resources. It is not used directly, but through the Drive hub.

PermissionPatchMethodBuilder
Updates a permission. This method supports patch semantics.

PermissionUpdateMethodBuilder
Updates a permission.

Property
A key-value pair that is either public or private to an application.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert (request|response)

get (response)

patch (request|response)

update (request|response)

PropertyDeleteMethodBuilder
Deletes a property.

PropertyGetMethodBuilder
Gets a property by its key.

PropertyInsertMethodBuilder
Adds a property to a file.

PropertyList
A collection of properties, key-value pairs that are either public or private to an application.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

PropertyListMethodBuilder
Lists a file's properties.

PropertyMethodsBuilder
A builder providing access to all methods supported on property resources. It is not used directly, but through the Drive hub.

PropertyPatchMethodBuilder
Updates a property. This method supports patch semantics.

PropertyUpdateMethodBuilder
Updates a property.

RealtimeGetMethodBuilder
Exports the contents of the Realtime API data model associated with this file as JSON.

RealtimeMethodsBuilder
A builder providing access to all methods supported on realtime resources. It is not used directly, but through the Drive hub.

RealtimeUpdateMethodBuilder
Overwrites the Realtime API data model associated with this file with the provided JSON data model.

ReplyDeleteMethodBuilder
Deletes a reply.

ReplyGetMethodBuilder
Gets a reply.

ReplyInsertMethodBuilder
Creates a new reply to the given comment.

ReplyListMethodBuilder
Lists all of the replies to a comment.

ReplyMethodsBuilder
A builder providing access to all methods supported on reply resources. It is not used directly, but through the Drive hub.

ReplyPatchMethodBuilder
Updates an existing reply. This method supports patch semantics.

ReplyUpdateMethodBuilder
Updates an existing reply.

Revision
A revision of a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete (none)

update (request|response)

patch (request|response)

list (none)

get (response)

RevisionDeleteMethodBuilder
Removes a revision.

RevisionGetMethodBuilder
Gets a specific revision.

RevisionList
A list of revisions of a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

RevisionListMethodBuilder
Lists a file's revisions.

RevisionMethodsBuilder
A builder providing access to all methods supported on revision resources. It is not used directly, but through the Drive hub.

RevisionPatchMethodBuilder
Updates a revision. This method supports patch semantics.

RevisionUpdateMethodBuilder
Updates a revision.

User
The JSON template for a user.

This type is not used in any activity, and only used as part of another schema.

UserPicture
The user's profile picture.

This type is not used in any activity, and only used as part of another schema.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.