Struct drive2::PermissionMethodsBuilder

pub struct PermissionMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on permission resources. It is not used directly, but through the Drive hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate drive2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `delete(...)`, `get(...)`, `get_id_for_email(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.permissions();

Methods

impl<'a, C, NC, A> PermissionMethodsBuilder<'a, C, NC, A>

fn delete(&self, file_id: &str, permission_id: &str) -> PermissionDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a permission from a file.

fn insert(&self, request: &Permission, file_id: &str) -> PermissionInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Inserts a permission for a file.

fn update(&self, request: &Permission, file_id: &str, permission_id: &str) -> PermissionUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a permission.

fn patch(&self, request: &Permission, file_id: &str, permission_id: &str) -> PermissionPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a permission. This method supports patch semantics.

fn list(&self, file_id: &str) -> PermissionListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists a file's permissions.

fn get(&self, file_id: &str, permission_id: &str) -> PermissionGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a permission by ID.

fn get_id_for_email(&self, email: &str) -> PermissionGetIdForEmailMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns the permission ID for an email address.